+++ powershell-awesome-framework/ReleaseNotes.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.0] - 2023.07.30
+
+### Added
+
+- `Get-PAFScriptBlockName`
+- show menu after invoke snippet script
+
+### Changed
+
+- snippet req. for PAF
+- no recurse on `Show-PAFSnippetMenu`
+
 ## [0.0.4] - 2023.07.28
 
 ### Added